pyproject.toml
[tool.poetry]
name = "contact_book"
version = "0.1.0"
description = "A Django application for managing contacts."
authors = ["Carol", "Ernesto"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
djangorestframework = "~3.14"
psycopg2-binary = "~2.9.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
